pyproject.toml
[project]
name = "arag"
version = "0.1.0"
description = "ARAG - Agentic Retrieval-Augmented Generation Framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ayanami0730", email = "ayanami0730@gmail.com"}
]
keywords = ["rag", "llm", "retrieval", "agent", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "tiktoken>=0.7.0",
    "transformers>=4.44.0",
    "datasets>=2.20.0",
    "accelerate>=0.33.0",
    "faiss-cpu>=1.8.0",
]

[project.optional-dependencies]
full = [
    "pyarrow>=12.0.0",
    "peft>=0.12.0",
    "trl>=0.10.1",
    "bitsandbytes>=0.43.1",
    "matplotlib>=3.8.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Ayanami0730/arag"
Repository = "https://github.com/Ayanami0730/arag"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arag"]

[tool.hatch.build.targets.sdist]
include = ["src/arag"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
